mighty.metadata: Manage 'CDISC' 'ADaM' Dataset Specifications in 'YAML' Format

Load, validate, and manipulate Clinical Data Interchange Standards Consortium ('CDISC') Analysis Data Model ('ADaM') dataset metadata stored as 'YAML' files. Metadata files are validated against a JSON schema. Provides functions to inspect and modify columns, parameters, and row-level operations within and across 'ADaM' domains. Designed for use with the 'mighty' framework.

See also

Author

Maintainer: Aksel Thomsen oath@novonordisk.com

Authors:

Aksel Thomsen oath@novonordisk.com

Ari Siggaard Knoph aikp@novonordisk.com

Matthew Phelps mewp@novonordisk.com

Giulia Pais egzp@novonordisk.com

Michael Hedegaard Thomsen mhzt@novonordisk.com

Other contributors:

Novo Nordisk A/S [copyright holder]